Feature/rent increment type (#305)
* decimal increment period calculation

* rent increment period modification

+++ rented-properties/src/main/java/org/egov/cpt/repository/PropertyRowMapper.java
@@ -52,7 +52,8 @@ public class PropertyRowMapper implements ResultSetExtractor<List<Property>> {
 						.currentOwner(rs.getString("current_owner")).floors(rs.getString("floors"))
 						.additionalDetails(rs.getString("additional_details")).address(address)
 						.rentIncrementPercentage(rs.getDouble("pd_rent_inc_pg"))
-						.rentIncrementPeriod(rs.getInt("pd_rent_inc_period")).interestRate(rs.getDouble("pd_int_rate"))
+						.rentIncrementPeriod(rs.getFloat("pd_rent_inc_period"))
+						.interestRate(rs.getDouble("pd_int_rate"))
 						.auditDetails(auditdetails).build();
 
 				currentProperty = Property.builder().id(propertyId).transitNumber(rs.getString("transit_number"))
